Introduce Money abstraction and use it in tests
We start with an empty Money struct for now and only add two factory methods for returning Dollar or Franc.

They don't return a common abstraction, as introducing a trait for it and use dynamic dispatching is not worth the effort.

The tests will rely on Money factory methods to get the structs, and as the type is inferred, no changes will be required after contracting Franc and Dollar into Money.